requirements.txt
pandas
scikit-learn
streamlit
python-dotenv
pytest
